players ints position

+++ Areserver/Actor.cs
@@ -4,9 +4,9 @@ namespace Areserver
 {
     public class Actor
     {
-        public float X { get; set; }
+        public int X { get; set; }
 
-        public float Y { get; set; }
+        public int Y { get; set; }
 
         public string Name { get; set; }
